Remove unused method in TimeAgoParser
